+++ src/p_016.rs
@@ -1,4 +1,4 @@
-fn explode_carry(mut digits: Vec<u64>) -> Vec<u64> {
+fn explode_carry(mut digits: Vec<i128>) -> Vec<i128> {
     let max_index = digits.len() - 1;
 
     for i in 0..digits.len() {
@@ -22,8 +22,8 @@ fn explode_carry(mut digits: Vec<u64>) -> Vec<u64> {
     digits
 }
 
-fn big_pow(base_n: u64, p: u64) -> Vec<u64> {
-    let mut digits: Vec<u64> = vec![1];
+fn big_pow(base_n: i128, p: i128) -> Vec<i128> {
+    let mut digits: Vec<i128> = vec![1];
 
     for _ in 0..p {
         for n in &mut digits {
@@ -37,6 +37,6 @@ fn big_pow(base_n: u64, p: u64) -> Vec<u64> {
 }
 
 #[allow(dead_code)]
-pub fn problem_016() -> u64 {
+pub fn problem_016() -> i128 {
     big_pow(2, 1000).iter().sum()
 }
