+++ src/p_019.rs
@@ -1,9 +1,9 @@
-fn is_leap_year(year: u64) -> bool {
+fn is_leap_year(year: i128) -> bool {
     year % 400 == 0 || (year % 4 == 0 && year % 100 != 0)
 }
 
 #[allow(dead_code)]
-pub fn problem_019() -> u64 {
+pub fn problem_019() -> i128 {
     let mut sunday_count = 0;
     let mut current_week_day = 1;
     let mut month_days: [usize; 12] = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
