+++ src/p_024.rs
@@ -1,13 +1,13 @@
 use itertools::Itertools;
 
 #[allow(dead_code)]
-pub fn problem_024() -> u64 {
+pub fn problem_024() -> i128 {
     let n = 9;
     let millionth_permutation = (0..=n).permutations(n + 1).nth(999_999).unwrap();
     millionth_permutation
         .iter()
         .map(|x| x.to_string())
         .join("")
-        .parse::<u64>()
+        .parse::<i128>()
         .unwrap()
 }
